Adds Github workflow for labelling PRs (#3404)
# Description

Given the increase in number of PRs to the codebase, it has become
harder to understand the type of changes and the corresponding team to
communicate with. This MR uses a GitHub bot to help us out with
categorizing the PRs.

More info: https://github.com/actions/labeler

## Type of change

- New feature (non-breaking change which adds functionality)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

---------

Signed-off-by: Mayank Mittal <12863862+Mayankm96@users.noreply.github.com>